Added separate file for coordinates
