+++ shm/Cargo.cpp
@@ -1,10 +1,11 @@
 #include "Cargo.hpp"
 
-Cargo::Cargo() : amount_(0), basePrice_(0) {}
-Cargo::Cargo(size_t amount, size_t basePrice) : amount_(amount), basePrice_(basePrice) {}
-
-
-
+Cargo::Cargo() : amount_(0), basePrice_(0)
+{
+}
+Cargo::Cargo(size_t amount, size_t basePrice) : amount_(amount), basePrice_(basePrice)
+{
+}
 
 Cargo& Cargo::operator+=(size_t amount)
 {
@@ -13,7 +14,8 @@ Cargo& Cargo::operator+=(size_t amount)
 }
 Cargo& Cargo::operator-=(size_t amount)
 {
-    if(amount_ >= amount) amount_ -= amount;
+    if (amount_ >= amount)
+        amount_ -= amount;
     return *this;
 }
