+++ shm/main.cpp
@@ -1,2 +1,4 @@
 
-int main() {}
+int main()
+{
+}
